requirements.txt
ccxt
pandas
numpy
requests
python-telegram-bot==20.3
scikit-learn
matplotlib
apscheduler
sqlalchemy
python-dotenv
psycopg2-binary
joblib